requirements.txt
fastapi
uvicorn
requests
google-cloud-storage
google-cloud-secret-manager
SQLAlchemy
pg8000
